Added ability to list users online (close #10)
Adds the ability to support a "list" command slack side in order to list players currently on the server. Requries Slash command integration on Slack's side, specifically token, channel name it'll be used on, and the actual command name ("list" by default). The URL should be the same as the outgoing webhook's and it should be a POST request.